data/pipelines/pipeline1_initial/requirements.txt
boto3
python-dotenv
pandas
pyarrow
numpy
scikit-learn
requests
